third_party/WebKit/LayoutTests/css3/device-adapt/opera/syntax-003-expected.txt


FAIL CSS Test: @viewport syntax - descriptor recovery assert_equals: expected 500 but got 800
